+++ docs/make.jl
@@ -14,7 +14,9 @@ makedocs(
     modules = [ClimaComms],
     pages = Any[
         "Home" => "index.md",
-        "Developing with `ClimaComms`" => "internals.md",
+        "Getting Started" => "getting_started.md",
+        "How-to Guide" => "howto.md",
+        "Design Philosophy" => "philosophy.md",
         "Logging" => "logging.md",
         "Frequently Asked Questions" => "faqs.md",
         "APIs" => "apis.md",